added learning agent to runner and populated table method

+++ src/DatabaseRecord.java
@@ -1,10 +1,10 @@
 public class DatabaseRecord {
     public int id;
-    public float value;
+    public double value;
     public int success;
     public int fail;
 
-    public DatabaseRecord(int id, float value, int success, int fail){
+    public DatabaseRecord(int id, double value, int success, int fail){
         this.id = id;
         this.value = value;
         this.success = success;
